+++ CHANGELOG.rst
@@ -19,6 +19,7 @@ Added
     - Sheet creation from templates using cubi-tk (#1068)
     - ``clean_sheet_dir_name()`` helper
     - iRODS delete requests for data objects and collections (#277, #1087, #1089, #1090)
+    - Allow per-project restriction of column config updates (#995)
 
 Changed
 -------